Fix Point dependencies
