+++ src/sailboat_main/sailboat_main/main_algo/Points.py
@@ -1,5 +1,6 @@
-import LatLongPoint
-import utm 
+import utm
+import math
+from sensor_msgs.msg import NavSatFix
 
 class UTMPoint():
     """
@@ -16,7 +17,7 @@ class UTMPoint():
         self.zone_number = zone_number
         self.zone_letter = zone_letter
 
-    def to_latlon(self) -> LatLongPoint:
+    def to_latlon(self) -> 'LatLongPoint':
         """
         Convert UTM coordinates to latitude and longitude.
         """
@@ -42,3 +43,23 @@ class UTMPoint():
 
     def __repr__(self):
         return f"UTMPoint(x={self.x}, y={self.y}, zone_number={self.zone_number}, zone_letter={self.zone_letter})"
+
+class LatLongPoint():
+    """
+    A class to represent a point in latitude and longitude coordinates.
+    """
+    latitude : float
+    longitude : float
+    def __init__(self, latitude, longitude):
+        self.latitude = latitude
+        self.longitude = longitude
+
+    def to_utm(self) -> 'UTMPoint':
+        """
+        Convert latitude and longitude to UTM coordinates.
+        """
+        x, y, zone_number, zone_letter = utm.from_latlon(self.latitude, self.longitude)
+        return UTMPoint(x, y, zone_number, zone_letter) 
+
+    def __repr__(self):
+        return f"latitude={self.latitude}, longitude={self.longitude}"
